update and add templates and forms for signup and login
