+++ hrMatching/mainapp/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from .forms import CustomUserCreationForm, CustomUserChangeForm
+from .forms.home import CustomUserCreationForm, CustomUserChangeForm
 from .models import User
