+++ hrMatching/mainapp/views/company.py
@@ -2,7 +2,7 @@ from django.contrib.auth import login
 from django.shortcuts import redirect
 from django.views.generic import CreateView
 
-from ..forms import CompanySignUpForm
+from ..forms.company import CompanySignUpForm
 from ..models import User
 
 class CompanySignUpView(CreateView):
