+++ hrMatching/mainapp/views/home.py
@@ -1,16 +1,16 @@
 from django.shortcuts import redirect, render
 from django.views.generic import TemplateView
-
+from django.contrib.auth.views import LoginView
+from ..forms.home import CustomLoginForm
 
 class SignUpView(TemplateView):
     template_name = 'registration/signup.html'
 
+class LoginView(LoginView):
+    form_class = CustomLoginForm
+    template_name = 'registration/login.html'
+
+
 
 def index(request):
-    """if request.user.is_authenticated:
-        if request.user.is_teacher:
-            return redirect('teachers:quiz_change_list')
-        else:
-            return redirect('students:quiz_list')
-    """
     return render(request, 'mainapp/index.html')
