+++ hrMatching/templates/registration/login.html
@@ -1,15 +1,15 @@
 {% extends "base.html" %}
 {% load crispy_forms_tags %}
 
-{% block title %}Company Sign Up{% endblock %}
+{% block title %} log in {% endblock %}
 
 {% block content %}
 <div class="container mt-5">
-    <h2>Company Sign Up</h2>
+    <h2>Log in</h2>
     <form method="post">
         {% csrf_token %}
         {{ form|crispy }}
-        <button type="submit" class="btn btn-primary">Sign Up</button>
+        <button type="submit" class="btn btn-primary">Log in</button>
     </form>
 </div>
 {% endblock %}
