+++ hrMatching/templates/registration/signup.html
@@ -1,8 +1,16 @@
 {% extends 'base.html' %}
 
 {% block content %}
-  <h2>Sign up for a free account</h2>
-  <p>Select below the type of account you want to create</p>
-  <a href="{% url 'company_signup' %}"  role="button">Company</a>
-  <a href="{% url 'employee_signup' %}" role="button">Employee</a>
+<div class="container mt-5">
+  <div class="row justify-content-center">
+    <div class="col-md-6 text-center">
+      <h2 class="mb-4">Sign up for a free account</h2>
+      <p class="mb-4">Select below the type of account you want to create</p>
+      <div class="d-grid gap-2">
+        <a href="{% url 'company_signup' %}" class="btn btn-primary btn-lg mb-2" role="button">Company</a>
+        <a href="{% url 'employee_signup' %}" class="btn btn-secondary btn-lg mb-2" role="button">Employee</a>
+      </div>
+    </div>
+  </div>
+</div>
 {% endblock %}
